+++ examples/basic/src/my-javascript-component-no-props/my-javascript-component-no-props.spec.js
@@ -1,0 +1,4 @@
+import React from 'react';
+import MyJavascriptComponentNoProps from 'my-javascript-component-no-props/my-javascript-component-no-props';
+
+export const SpecMyJavascriptComponentNoPropsDefault = () => <MyJavascriptComponentNoProps />;