+++ src/app-global-styles.js
@@ -1,0 +1,32 @@
+import { createGlobalStyle } from 'styled-components';
+import { normalize } from 'styled-normalize';
+
+const GlobalStyle = createGlobalStyle`
+    *,
+    *::before,
+    *::after {
+        box-sizing: border-box;
+    }
+
+    body {
+        color: #6a6a6a;
+    }
+
+    h1 {
+        font-weight: 400;
+    }
+
+    code {
+        font-family: Consolas, "Liberation Mono", Menlo, monospace;
+    }
+
+    .styleguide {
+        background: #fff;
+        display: grid;
+        grid-template-columns: 300px auto;
+    }
+
+    ${normalize}
+`;
+
+export default GlobalStyle;