+++ src/components/error-boundary/error-boundary.jsx
@@ -1,6 +1,28 @@
 import React from 'react';
+import styled from 'styled-components';
 
-import './error-boundary.scss';
+const ErrorBoundaryBlock = styled.section`
+    font-family: monospace;
+    line-height: 1.5;
+    padding: 16px;
+    background: rgba(222, 23, 56, 0.2);
+    color: rgba(222, 23, 56, 1);
+`;
+
+const ErrorBoundaryTitle = styled.h3`
+    margin: 0;
+`;
+
+const ErrorBoundaryShowMore = styled.p`
+    margin: 10px 0 0;
+    cursor: pointer;
+    text-decoration: underline;
+`;
+
+const ErrorBoundaryText = styled.p`
+    margin: 10px 0 0;
+    white-space: pre-wrap;
+`;
 
 class ErrorBoundary extends React.Component {
     constructor(props) {
@@ -19,27 +41,24 @@ class ErrorBoundary extends React.Component {
     render() {
         if (this.state.error) {
             return (
-                <section className="error-boundary">
-                    <h3 className="error-boundary__title">{this.state.error.toString()}</h3>
+                <ErrorBoundaryBlock>
+                    <ErrorBoundaryTitle>{this.state.error.toString()}</ErrorBoundaryTitle>
+
                     {this.state.showStack ? (
                         <React.Fragment>
-                            <p
-                                className="error-boundary__show-more"
+                            <ErrorBoundaryShowMore
                                 onClick={() => this.setState({ showStack: false })}
                             >
                                 Hide error stack
-                            </p>
-                            <p className="error-boundary__text">{this.state.error.stack}</p>
+                            </ErrorBoundaryShowMore>
+                            <ErrorBoundaryText>{this.state.error.stack}</ErrorBoundaryText>
                         </React.Fragment>
                     ) : (
-                        <p
-                            className="error-boundary__show-more"
-                            onClick={() => this.setState({ showStack: true })}
-                        >
+                        <ErrorBoundaryShowMore onClick={() => this.setState({ showStack: true })}>
                             Show error stack
-                        </p>
+                        </ErrorBoundaryShowMore>
                     )}
-                </section>
+                </ErrorBoundaryBlock>
             );
         }